### 2019-09-03 #1702 @whedon
[PRE REVIEW]: PYCV: a PLUMED 2 Module Enabling the Rapid Prototyping of Collective Variables in Python
**Submitting author:** @tonigi (<a href="http://orcid.org/0000-0001-6449-0596">Toni Giorgino</a>)
**Repository:** <a href="https://github.com/giorginolab/plumed2-pycv/" target ="_blank">https://github.com/giorginolab/plumed2-pycv/</a>
**Version:** v1.1
