### 2019-09-03 #1702 @arfon
@tonigi - thanks for this submission to JOSS. Could you please identify a few potential reviewers for this submission. In particular, if at all possible I'd like to have one of the PLUMED 2 core team involved in this review.
