### 2019-09-03 9cce176
Travis
commit 657b07022a9c2bee8163c10436a65b47ca7b3a6a
Author: Toni <toni.giorgino@gmail.com>
Date:   Tue Sep 3 18:47:43 2019 +0200

    *** empty log message ***

commit b0f7b301ef25bb06385d108061a7a1643a2f897d
Author: Toni G <tonigi@users.noreply.github.com>
Date:   Tue Sep 3 18:33:58 2019 +0200

    Update install.pycv.prereqs

commit 2d62d99a45d2d159bb6d55f21600a21672b4ed14
Author: Toni <toni.giorgino@gmail.com>
Date:   Tue Sep 3 18:26:35 2019 +0200

    osx

commit 3767591d3bf0313cf2b2f4aa7568f6c5f21e3825
Author: Toni G <tonigi@users.noreply.github.com>
Date:   Tue Sep 3 18:08:04 2019 +02
[…]
