+++ index.md
@@ -1,0 +1,211 @@
+---
+layout: default
+---
+
+The PYCV module for PLUMED 2
+====================================
+
+Enable the implementation of collective variables (CVs) and
+functions in the Python language.
+
+Advantages of using PYCV over standard development of CVs in C++ are:
+ 1. functions may be prototyped in  high-level code, using
+    extensive mathematical libraries, and no boilerplate;
+ 2. just-in-time compilation
+    occurs transparently: there are no compilation and link delays
+    for code changes;
+ 3. CVs may be automatically differentiated in common cases.
+
+The code is organized as a standard PLUMED module: development occurs
+in a [fork of the original
+repository](https://github.com/giorginolab/plumed2-pycv/tree/v2.5.2-pycv/src/pycv). All
+code is in the `src/pycv` and `regtest/pycv` directories.
+
+
+Documentation
+------------------------------------
+
+The PYCV module defines the following actions:
+
+ * `PYTHONCV`, to implement single- and multi-component CVs in Python;
+ * `PYTHONFUNCTION`, to implement arbitrary functions.
+
+In both cases, a Python interpreter is first started; the Python file
+indicated in the `IMPORT=` keyword is then imported; from it, an
+user-chosen function (`FUNCTION=`) is called to perform the
+computations at each timestep. Modules can be shared for multiple
+functions, and contain one-time initialization.
+
+Transparent auto-differentiation, JIT compilation, and vectorization
+are available through Google's [JAX
+library](https://github.com/google/jax) (recommended).
+
+Please see the generated documentation and regression tests.
+
+
+
+
+Prerequisites
+------------------------------------
+
+Python 3 is required. Under OSX, [Homebrew](https://brew.sh)'s Python
+3 package works well. The Conda version is problematic (slow
+compilation, linking failures).
+
+You will also need to install `numpy`, either via `pip3 install
+numpy` or your distribution's packages.
+
+Automatic differentiation examples require the JAX library: install
+it with `pip3 install jax jaxlib`. 
+
+
+
+Installation
+------------------------------------
+
+Follow the usual PLUMED 2 configuration procedure:
+
+```bash
+./configure --enable-modules=+pycv 
+make -j4
+make install   # optional
+```
+
+Please inspect the configure messages to see if any missing dependency
+prevents PYCV from actually being enabled. Verify the successful installation
+with e.g.
+
+    ./src/lib/plumed-static manual --action PYTHONCV
+
+which should return the manual for `PYTHONCV`. (The `plumed-static`
+executable should work even without installation.)
+
+(It is also possible to compile the module as a `LOAD`-able dynamic
+object.  Once in the `src/pycv` directory, issue `make PYCV.so`
+(`PYCV.dylib` under OSX). The compilation step *should* pick
+Python-specific flags as long as the correct `python3-config`
+executable is in your path.)
+
+
+Quickstart
+------------------------------------
+
+Here's a quick example to whet your appetite, following the regression test `rt-jax2`.
+
+    export PLUMED_PROGRAM_NAME=$PWD/src/lib/plumed-static 
+    make -C regtest/pycv/rt-jax2
+
+The files in `regtest/pycv/rt-jax2` are reproduced here:
+
+**File plumed.dat**
+
+```
+cv1:  PYTHONCV ATOMS=1,4,3 IMPORT=jaxcv FUNCTION=cv1
+```
+
+**File jaxcv.py**
+
+```py
+# Import the JAX library
[…]
