+++ _config.yml
@@ -7,11 +7,12 @@ author:
   github: tonigi
 
 title: The PYCV module for PLUMED 2
-description: PYCV enables PLUMED2 Collective Variables (CVs) and arbitrary functions to be defined and auto-differentiated in the Python language. 
-logo: /images/logo.svg
+description: Enable the implementation of collective variables (CVs) and functions in the Python language.
+logo: /images/logo.png
 repository: giorginolab/plumed2-pycv
 
-show_downloads: true
+
+show_downloads: false
 markdown: kramdown
 google_analytics: nil
 theme: jekyll-theme-minimal
