+++ _config.yml
@@ -10,7 +10,7 @@ title: The PYCV module for PLUMED 2
 description: Enable the implementation of collective variables (CVs) and functions in the Python language.
 logo: /images/logo.png
 repository: giorginolab/plumed2-pycv
-
+url: https://giorginolab.github.io/plumed2-pycv/
 
 show_downloads: false
 markdown: kramdown
