### 2019-09-20 d906dba
Set theme jekyll-theme-minimal
