### 2019-09-30 #1702 @carlocamilloni
Yes this is a good time to me

> On 30 Sep 2019, at 15:35, Toni G <notifications@github.com> wrote:
>
> Perhaps @carlocamilloni <https://github.com/carlocamilloni> could be so kind to give it an eye?
>
