### 2019-09-30 #1702 @tonigi
To summarize:
* Fork of Plumed 2.5 with the `pycv` module: https://github.com/giorginolab/plumed2-pycv/
* Web page: https://giorginolab.github.io/plumed2-pycv/
* Paper (auto-generated by the editorial bot) https://github.com/openjournals/joss-papers/blob/joss.01702/joss.01702/10.21105.joss.01702.pdf
* Info on the review process: https://joss.readthedocs.io/en/latest/review_criteria.html  (will take place in a future issue, as far as I know)

Thanks!
