### 2019-10-01 42babed
readding joss paper in default branch
