+++ src/pycv/joss-paper/paper.bib
@@ -1,0 +1,248 @@
+
+@article{tribello_plumed_2014,
+	title = {{PLUMED} 2: {New} feathers for an old bird},
+	volume = {185},
+	issn = {0010-4655},
+	shorttitle = {{PLUMED} 2},
+	url = {http://www.sciencedirect.com/science/article/pii/S0010465513003196},
+	doi = {10.1016/j.cpc.2013.09.018},
+	abstract = {Abstract
+Enhancing sampling and analyzing simulations are central issues in molecular simulation. Recently, we introduced PLUMED, an open-source plug-in that provides some of the most popular molecular dynamics (MD) codes with implementations of a variety of different enhanced sampling algorithms and collective variables (CVs). The rapid changes in this field, in particular new directions in enhanced sampling and dimensionality reduction together with new hardware, require a code that is more flexible and more efficient. We therefore present PLUMED 2 here—a complete rewrite of the code in an object-oriented programming language (C++). This new version introduces greater flexibility and greater modularity, which both extends its core capabilities and makes it far easier to add new methods and CVs. It also has a simpler interface with the MD engines and provides a single software library containing both tools and core facilities. Ultimately, the new code better serves the ever-growing community of users and contributors in coping with the new challenges arising in the field.
+Program summary
+Program title: PLUMED 2
+
+Catalogue identifier: AEEE\_v2\_0
+
+Program summary URL:http://cpc.cs.qub.ac.uk/summaries/AEEE\_v2\_0.html
+
+Program obtainable from: CPC Program Library, Queen’s University, Belfast, N. Ireland
+
+Licensing provisions: Yes
+
+No. of lines in distributed program, including test data, etc.: 700646
+
+No. of bytes in distributed program, including test data, etc.: 6618136
+
+Distribution format: tar.gz
+
+Programming language: ANSI-C++.
+
+Computer: Any computer capable of running an executable produced by a C++ compiler.
+
+Operating system: Linux operating system, Unix OSs.
+
+Has the code been vectorized or parallelized?: Yes, parallelized using MPI.
+
+RAM: Depends on the number of atoms, the method chosen and the collective variables used.
+
+Classification: 3, 7.7, 23.
+
+Catalogue identifier of previous version: AEEE\_v1\_0.
+
+Journal reference of previous version: Comput. Phys. Comm. 180 (2009) 1961.
+
+External routines: GNU libmatheval, Lapack, Blas, MPI.
+
+Does the new version supersede the previous version?: This version supersedes the previous version for the most part. There are a small number of very specific situations where the previous version is better, due to performance or to non-ported features. We are actively working on porting these last few features into the new code.
+
+Nature of problem:
+
+Calculation of free-energy surfaces for molecular systems of interest in biology, chemistry and materials science, on the fly and a posteriori analysis of molecular dynamics trajectories using advanced collective variables.
+
+Solution method:
+
+Implementations of various collective variables and enhanced sampling techniques.
+
+Reasons for new version:
+
+The old version was difficult to maintain and its design was not as flexible as this new version. This lack of flexibility made it difficult to implement a number of novel methods that have emerged since the release of the original code.
+
+Summary of revisions:
+
[…]
