+++ src/pycv/joss-paper/paper.md
@@ -1,0 +1,172 @@
+---
+title: 'PYCV: a PLUMED 2 Module Enabling the Rapid Prototyping of Collective Variables in Python'
+tags:
+  - Python
+  - PLUMED
+  - molecular dynamics
+  - collective variables
+  - biased sampling
+  - gradient
+authors:
+  - name: Toni Giorgino
+    orcid: 0000-0001-6449-0596
+    affiliation: 1
+affiliations:
+ - name: Institute of Biophysics (IBF-CNR), National Research Council of Italy
+   index: 1
+date: 15 August 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+Collective variables (CVs) are functions of the coordinates of
+particles in a molecular system. The choice of CV is
+crucial to capture relevant degrees of freedom of the model being
+simulated [@barducci_metadynamics_2011]. This is especially important
+when employing *biased sampling* techniques such as umbrella sampling
+or metadynamics [@torrie_nonphysical_1977; @laio_escaping_2002], which
+apply generalized forces to CVs to enhance the sampling of events
+otherwise not observable by direct simulation.   CVs may be
+simple geometrical observables (distances, angles, torsions, etc.),
+but are often more complex functions designed to capture structural
+determinants, such as tertiary and quaternary structure of proteins,
+experimental observables, crystal symmetries, etc. [@branduardi_b_2007;
+@bonomi_integrative_2017; @pipolo_navigating_2017].
+
+Iterative development of CVs therefore occupies a large fraction of
+the effort in the exploration of molecular systems. On the one hand,
+the task has been largely facilitated by biasing libraries such as
+PLUMED [@tribello_plumed_2014], which provides pre-defined functions and a
+*lingua franca* to express CV combinations, atom groups and biasing
+schemes, and its active community [@the_plumed_consortium_promoting_2019].
+However, users willing to explore CVs beyond the pre-defined
+ones have to implement them in C++, together with the corresponding
+(often cumbersome) derivatives [@giorgino_how_2018]. Compiled code is
+unwieldy for iterative analysis, because it is relatively low-level,
+error-prone, and inconvenient in exploratory stages.
+
+This paper introduces **PYCV**, a module for the PLUMED 2 library
+which enables users to define CVs and arbitrary functions in the
+Python language.  CV implementations may thus be modified and tested
+independently of the main code, with essentially no "test latency".
+Of note, coordinates are processed as `numpy` arrays, making it
+convenient to leverage the vast set of linear algebra and numerical
+algorithms provided by `numpy`, `scipy`, and many other open-source
+modules. Furthermore, just-in-time compilation and reverse-mode
+automatic differentiation are easily accessible using Google's JAX
+library.
+
+
+# Usage
+
+The **PYCV** module registers itself with PLUMED to provide the
+following actions:
+
+ * `PYTHONCV`, to implement single- and multi-component CVs;
+ * `PYTHONFUNCTION`, to implement arbitrary functions.
+
+The actions are documented in the respective inline manuals (e.g.,
+`plumed manual --action PYTHONCV`).  In both cases, an interpreter is
+first started; the Python module indicated in the `IMPORT=` keyword is
+then loaded; from it, an user-chosen function (`FUNCTION=`) is called
+to perform the computations at each timestep. Modules can contain
+multiple functions and one-time initialization.
+
+
+
+# Example
+
+A self-explanatory example is provided for illustration below. It is
+equivalent to the *radius of curvature* example shown in
[…]
