### 2019-10-01 #1773 @tonigi
Perhaps stuck because the paper is not in the default branch. Re-committed there and retrying.
