+++ src/pycv/joss-paper/paper.bib
@@ -130,14 +130,14 @@ Program Title: Practical approaches to the differentiation of collective variabl
 	title = {The calculation of the potential of mean force using computer simulations},
 	volume = {91},
 	url = {Roux1995.pdf},
-	doi = {http://dx.doi.org/DOI: 10.1016/0010-4655(95)00053-I},
+	doi = {10.1016/0010-4655(95)00053-I},
 	abstract = {The problem of unbiasing and combining the results of umbrella sampling calculations is reviewed. The weighted histogram analysis method (WHAM) of S. Kumar et al. (J. Comp. Chem. 13 (1992) 1011) is described and compared with other approaches. The method is illustrated with molecular dynamics simulations of the alanine dipeptide for one-and two-dimensional free energy surfaces. The results show that the WHAM approach simplifies considerably the task of recombining the various windows in complex systems.},
 	number = {1-3},
 	journal = {Computer Physics Communications},
 	author = {Roux, Benoît},
 	year = {1995},
 	keywords = {Computer simulations},
-	pages = {275 --282},
+	pages = {275--282},
 	annote = {Added to JabRef: 2009.01.09},
 	annote = {Owner: toni}
 }
