+++ README.md
@@ -5,12 +5,14 @@ Toni Giorgino
 
 This repository contains the **PYCV** module of PLUMED. 
 
+**Important.** Following PLUMED's [contribution philosophy](https://www.plumed.org/doc-v2.5/developer-doc/html/_how_to_contribute_to_plumed.html), this repository is a fork of the official PLUMED2 repository taken at release v2.5.2. The contributed code is **entirely** contained in the directories [`src/pycv`](src/pycv) and [`regtest/pycv`](regtest/pycv).  Minor modifications were made to the `configure` script to simplify installation.
+
+
 PYCV is a module for the [PLUMED 2 library](https://www.plumed.org) which enables users to define Collective Variables (CVs) and arbitrary functions in the Python language. CV implementations may thus be modified and tested independently of the main code, with essentially no "test latency". Of note, coordinates are processed as numpy arrays, making it convenient to leverage the vast set of linear algebra and numerical algorithms provided by numpy, scipy, and many other open-source modules. Furthermore, just-in-time compilation and reverse-mode automatic differentiation are easily accessible using Google's [JAX library](https://github.com/google/jax).
 
 The module's documentation and build instructions are in the [README.md](src/pycv/README.md) file contained in the module's directory. The actions contributed by this module are [self-documenting](https://www.plumed.org/doc-v2.5/developer-doc/html/using_doxygen.html); the corresponding text can be accessed e.g. by `plumed manual --action PYTHONCV`, or in the docs generated by `make docs`.  PYCV has a homepage at [giorginolab.github.io/plumed2-pycv](https://giorginolab.github.io/plumed2-pycv).
 
 
-**Important.** Following PLUMED's [contribution philosophy](https://www.plumed.org/doc-v2.5/developer-doc/html/_how_to_contribute_to_plumed.html), this repository is a fork of the official PLUMED2 repository taken at release v2.5.2. The contributed code is **entirely** contained in the directories [`src/pycv`](src/pycv) and [`regtest/pycv`](regtest/pycv).  Minor modifications were made to the `configure` script to simplify installation.
 
 Please report bugs & ideas via this repository's *Issues*. PLUMED's original README is [here](README.md.upstream). 
 
