### 2019-10-07 giorginolab/plumed2-pycv#1 @amritagos
[JOSS] Minor Paper Formatting
Part of https://github.com/openjournals/joss-reviews/issues/1773. The phrase changes are suggestions, and similar grammatical updates will be considered. Please fix the following minor issues with the manuscript:
## Paragraph I
![image](https://user-images.githubusercontent.com/32496815/66309529-7f490900-e927-11e9-97fa-3b12e064e55e.png)
Double spacing in the first paragraph.
## Paragraph II
- "occupies a large fraction of the effort in the exploration.." -> "accounts for most of the efforts associated with the exploration..."
- "biasing libraries" --> Intent unclear. PLUMED is a library which includes enhanced sampling techniques for steered or 'biased' MD, but clarification is required.
- 'which provides' --> Unclear, sentence structure makes it seem like it refers to the "biasing libraries" when it appears to have been intended for PLUMED (suggestion: refactor sentence)
## Code
![image](https://user-images.githubusercontent.com/32496815/66310125-0480ed80-e929-11e9-8916-4fc5fc0eeb11.png)
- CV function is actually called
## Conclusions
- Since these are important conclusions, it might be better to spread them out as a numbered list, one per line.
