### 2019-10-07 giorginolab/plumed2-pycv#2 @amritagos
[JOSS] Contributing Guidelines
Part of https://github.com/openjournals/joss-reviews/issues/1773. The contributing guidelines should be in their own heading at least, and possibly in a file at the root of the repository, labelled `CONTRIBUTING`, with more details.
