### 2019-10-07 #1773 @amritagos
@tonigi The commit style is not clear at all.
![image](https://user-images.githubusercontent.com/32496815/66310730-8faeb300-e92a-11e9-8b91-aa40bb3abb9c.png)
However it is too late to change that so do keep it in mind for future work. Once the issues have been resolved I will be happy to complete my review.
