### 2019-10-07 b5bb113
reword paper as per reviews @amritagos 's suggestions
