### 2019-10-08 giorginolab/plumed2-pycv#1 @tonigi
Thanks! Changes are in b5bb113a0bf20709c97df1998689072951b17e6a , and in particular

* Double space changed in source: However it may not reflect in the output because the spacing is in control of the typesetting engine. (Latex's spacing rules for intersentence space)
* Paragraph II: rewritten
* Code: reworded as "PLUMED will call the following function"
* Conclusions: itemized as requested.
