### 2019-10-08 giorginolab/plumed2-pycv#2 @tonigi
Introduced the heading in b0855a40e1c96c8c6e0174e5f61489bd81f78e68 .
