### 2019-10-08 giorginolab/plumed2-pycv#2 @amritagos
That will be sufficient, thanks.
