### 2019-10-10 #1773 @carlocamilloni
@arfon - @tonigi contribution is well coded and documented and provide a smart tool for prototyping new code for plumed. Python is definitely the way to go. The paper is self-contained and provides all what one needs to immediately use it.
